Title Menu Improvements INC OPTIONS
Added the ability to set options at the main menu, maybe later we can have the copyright flag be apart of it?

Added a fade in/out between the disclaimer/warning and the title, reduced the time on a couple of 30 frame waits by 10.

Made a few minor changes to the opening cutscene.