FOUNDATION FOR ANY CARD GAME
Check the GO FISH map for more details.
Basically I spent over 5 hours implementing the abilty for the game to pick a random card from a standard 52 card deck. It will never pick the same card twice and when the deck is empty it will give a warning.

I'm going to sleep now because holy fucking shit.