Work on Enemies + Misc
Added colored versions of two enemies since they're unable to be hue shifted.
Tweaks to alpha enemy stats + more formations for them.
Tweaks to 1993 party stats.
Changed enemy flee to not have a 100% success rate, now 60% success rate.
Tweaked sell price for 1993 items.
Added a new camera sprite and plonked a few of them on the overworld.

That's it for now!